backend/requirements.txt
Django==4.2.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0
Faker==25.2.0